+++ weevent-broker/src/main/java/com/webank/weevent/broker/fisco/FiscoBcosBroker4Producer.java
@@ -3,6 +3,7 @@ package com.webank.weevent.broker.fisco;
 import java.nio.charset.StandardCharsets;
 import java.util.concurrent.CompletableFuture;
 
+import com.webank.weevent.broker.fisco.util.DataTypeUtils;
 import com.webank.weevent.broker.fisco.util.ParamCheckUtils;
 import com.webank.weevent.broker.fisco.web3sdk.FiscoBcosDelegate;
 import com.webank.weevent.broker.plugin.IProducer;
@@ -10,7 +11,6 @@ import com.webank.weevent.sdk.BrokerException;
 import com.webank.weevent.sdk.SendResult;
 import com.webank.weevent.sdk.WeEvent;
 
-import com.alibaba.fastjson.JSON;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -42,6 +42,6 @@ public class FiscoBcosBroker4Producer extends FiscoBcosTopicAdmin implements IPr
         return fiscoBcosDelegate.publishEvent(event.getTopic(),
                 Long.parseLong(groupId),
                 new String(event.getContent(), StandardCharsets.UTF_8),
-                JSON.toJSONString(event.getExtensions()));
+                DataTypeUtils.object2Json(event.getExtensions()));
     }
 }
