+++ weevent-broker/src/main/resources/fisco.properties
@@ -1,7 +1,7 @@
 #fisco version
 version=2.0
 #fisco channel nodes, like 127.0.0.1:20200
-nodes=127.0.0.1:20200
+nodes=10.107.105.228:20200
 orgid=fisco
 #god account
 account=bcec428d5205abe0f0cc8a734083908d9eb8563e31f943d760786edf42ad67dd
