+++ weevent-governance/build.gradle
@@ -24,9 +24,8 @@ dependencies {
     implementation("org.jsoup:jsoup:${jsoupVersion}")
 
     //jacksonCore
-    implementation("com.fasterxml.jackson.core:jackson-databind:${jackSonVerion}")
-    implementation("com.fasterxml.jackson.module:jackson-module-jaxb-annotations:${jackSonVerion}")
-    implementation("com.fasterxml.jackson.core:jackson-core:${jackSonVerion}")
+    implementation("com.fasterxml.jackson.core:jackson-databind:${jackSonVersion}")
+    implementation("com.fasterxml.jackson.core:jackson-core:${jackSonVersion}")
 
     // forward backend
     implementation("org.apache.httpcomponents:httpclient:${httpclientVersion}")
