+++ weevent-robust/build.gradle
@@ -10,9 +10,8 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-log4j2:${springBoot2Version}")
     // avoid "Log4j appears to be running in a Servlet environment, but there's no log4j-web module available."
     //jacksonCore
-    implementation("com.fasterxml.jackson.core:jackson-databind:${jackSonVerion}")
-    implementation("com.fasterxml.jackson.module:jackson-module-jaxb-annotations:${jackSonVerion}")
-    implementation("com.fasterxml.jackson.core:jackson-core:${jackSonVerion}")
+    implementation("com.fasterxml.jackson.core:jackson-databind:${jackSonVersion}")
+    implementation("com.fasterxml.jackson.core:jackson-core:${jackSonVersion}")
 
     implementation("commons-io:commons-io:${apacheCommonsVersion.commonsIo}")
 }
